+++ README.md
@@ -1,69 +1,75 @@
 # tennis-competition
 
-[![Release](https://img.shields.io/github/v/release/julionevadod/tennis-competition)](https://img.shields.io/github/v/release/julionevadod/tennis-competition)
-[![Build status](https://img.shields.io/github/actions/workflow/status/julionevadod/tennis-competition/main.yml?branch=main)](https://github.com/julionevadod/tennis-competition/actions/workflows/main.yml?query=branch%3Amain)
-[![codecov](https://codecov.io/gh/julionevadod/tennis-competition/branch/main/graph/badge.svg)](https://codecov.io/gh/julionevadod/tennis-competition)
-[![Commit activity](https://img.shields.io/github/commit-activity/m/julionevadod/tennis-competition)](https://img.shields.io/github/commit-activity/m/julionevadod/tennis-competition)
-[![License](https://img.shields.io/github/license/julionevadod/tennis-competition)](https://img.shields.io/github/license/julionevadod/tennis-competition)
+Current projects trains a two DDPG agents to solve tennis competition unity environment.
+![alt text](tennis-agents-match.gif)
 
-Project to solve Tennis unity environment
+## Project Details
 
-- **Github repository**: <https://github.com/julionevadod/tennis-competition/>
-- **Documentation** <https://julionevadod.github.io/tennis-competition/>
+### The Environment
 
-## Getting started with your project
+States are defined in a 8-dimensional space. Action space size is 2-dimensional an continuous. The two dimensions refer to movement towards or away from the net and jumping.
+The two DDPG agents are trained in an adversarial setting. The reward of each episode is taken from the maximum reward among both agents. Each agent has isolated views of the space state and take their own actions (independent networks) with their own replay buffers.
 
-### 1. Create a New Repository
+### The task
 
-First, create a repository on GitHub with the same name as this project, and then run the following commands:
+The task is episodic, meaning that it has a defined end state (marked by done flag coming from environment). An eposide is considered terminated when an agent lets the ball hit the ground OR kicks it out of bounds. The task is considered to be solved when one of the agents achieves an average reward of +0.5 over 100 consecutive episodes.
 
-```bash
-git init -b main
-git add .
-git commit -m "init commit"
-git remote add origin git@github.com:julionevadod/tennis-competition.git
-git push -u origin main
-```
+## Getting started
 
-### 2. Set Up Your Development Environment
+. Fork the `tennis-competition` repo on GitHub.
 
-Then, install the environment and the pre-commit hooks with
+1. Clone your fork locally:
 
 ```bash
-make install
+cd <directory_in_which_repo_should_be_created>
+git clone https://github.com/julionevadod/tennis-competition.git
 ```
 
-This will also generate your `uv.lock` file
-
-### 3. Run the pre-commit hooks
-
-Initially, the CI/CD pipeline might be failing due to formatting issues. To resolve those run:
+2. Now we need to install the environment. Navigate into the directory
 
 ```bash
-uv run pre-commit run -a
+cd tennis-competition
 ```
 
-### 4. Commit the changes
-
-Lastly, commit the changes made by the two steps above to your repository.
+3. Then, install the environment with:
 
 ```bash
-git add .
-git commit -m 'Fix formatting issues'
-git push origin main
+uv sync
 ```
 
-You are now ready to start development on your project!
-The CI/CD pipeline will be triggered when you open a pull request, merge to main, or when you create a new release.
+4. Place Tennis.app from course resources inside **env** folder.
 
-To finalize the set-up for publishing to PyPI, see [here](https://fpgmaas.github.io/cookiecutter-uv/features/publishing/#set-up-for-pypi).
-For activating the automatic documentation with MkDocs, see [here](https://fpgmaas.github.io/cookiecutter-uv/features/mkdocs/#enabling-the-documentation-on-github).
-To enable the code coverage reports, see [here](https://fpgmaas.github.io/cookiecutter-uv/features/codecov/).
+## Instructions
 
-## Releasing a new version
+Once environment has been activated, training can be run from command line:
 
+```bash
+uv run train.py
+```
 
+Running the module as it is will run the training loop with default parameters. These default parameters produce an agent that solves the environment. However, agent hyperparameters can be configured by means of runtime arguments:
 
----
+#### `-i`, `--iterations`
 
-Repository initiated with [fpgmaas/cookiecutter-uv](https://github.com/fpgmaas/cookiecutter-uv).
+- **Description**: Number of environment steps to train for.
+- **Default**: Value from `config["DEFAULT"]["ITERATIONS"]`.
+
+#### `-b`, `--batch_size`
+
+- **Description**: Batch size for learning steps.
+- **Default**: Value from `config["DEFAULT"]["BATCH_SIZE"]`.
+
+#### `-g`, `--gamma`
+
+- **Description**: Discount rate used in reinforcement learning.
+- **Default**: Value from `config["DEFAULT"]["GAMMA"]`.
+
+#### `-a`, `--learning_rate_actor`
+
+- **Description**: Learning rate for the Actor optimizer.
+- **Default**: Value from `config["DEFAULT"]["LEARNING_RATE_ACTOR"]`.
+
+#### `-c`, `--learning_rate_critic`
+
+- **Description**: Learning rate for the Critic optimizer.
+- **Default**: Value from `config["DEFAULT"]["LEARNING_RATE_CRITIC"]`.
