--- conf/conf.ini
@@ -1,7 +1,0 @@
-[DEFAULT]
-ITERATIONS=1800
-BATCH_SIZE=128
-GAMMA=0.9
-LEARNING_RATE_ACTOR=1e-3
-LEARNING_RATE_CRITIC=1e-3
-SAVE_PATH=checkpoints/
