+++ pyproject.toml
@@ -18,6 +18,7 @@ classifiers = [
     "Topic :: Software Development :: Libraries :: Python Modules",
 ]
 dependencies = [
+    "optuna>=4.2.1",
     "unityagents",
 ]
