+++ tennis_competition/conf/conf.ini
@@ -1,0 +1,10 @@
+[DEFAULT]
+ITERATIONS=18000
+BATCH_SIZE=52
+GAMMA=0.93
+LEARNING_RATE_ACTOR=1e-4
+LEARNING_RATE_CRITIC=1e-4
+UPDATE_EVERY=2
+SIGMA=1
+SIGMA_DECAY=0.95
+SAVE_PATH=checkpoints/
